Get rid of the problem connected to simultaneous work of DevTools extension and vanilla DevTools
