+++ package.json
@@ -87,6 +87,8 @@
     "react-redux": "4.4.5",
     "react-router": "2.6.1",
     "react-router-redux": "4.0.5",
-    "redux": "3.5.2"
+    "redux": "3.5.2",
+    "redux-logger": "2.6.1",
+    "redux-promise": "0.5.3"
   }
 }
