+++ src/js/components/Root.js
@@ -25,7 +25,7 @@ const Root = ({ store, history }) => {
       <Provider store={store}>
         <div>
           <Router history={history} routes={routes} />
-          <DevTools />
+          {!window.devToolsExtension ? <DevTools /> : null}
         </div>
       </Provider>
     );
