+++ src/js/store/configureStore.js
@@ -1,47 +1,6 @@
-import { createStore, applyMiddleware, compose } from 'redux';
-import rootReducer from '../reducers';
-import { persistState } from 'redux-devtools';
-
-export default function configureStore(initialState) {
-
-  let middleware = applyMiddleware();
-  let enhancer;
-
-  if (process.env.NODE_ENV !== 'production') {
-
-    let middlewares = [require('redux-immutable-state-invariant')()];
-    middleware = applyMiddleware(...middlewares);
-
-    let getDebugSessionKey = function () {
-      // By default we try to read the key from ?debug_session=<key> in the address bar
-      const matches = window.location.href.match(/[?&]debug_session=([^&]+)\b/);
-      return (matches && matches.length) ? matches[1] : null;
-    };
-
-    enhancer = compose(
-
-      // Middleware we want to use in development
-      middleware,
-      window.devToolsExtension ?
-        window.devToolsExtension() :
-        require('../containers/DevTools').default.instrument(),
-
-      // Optional. Lets you write ?debug_session=<key> in address bar to persist debug sessions
-      persistState(getDebugSessionKey())
-    );
-  } else {
-    enhancer = compose(middleware);
-  }
-
-  const store = createStore(rootReducer, initialState, enhancer);
-
-  // Enable Webpack hot module replacement for reducers
-  if (module.hot) {
-    module.hot.accept('../reducers', () => {
-      const nextReducer = require('../reducers').default;
-      store.replaceReducer(nextReducer);
-    });
-  }
-
-  return store;
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./configureStore.production');
+}
+else {
+  module.exports = require('./configureStore.development');
 }
